Finished the 2015 setup
Registration form has an additional field added to record the
year of study. There was a bug in the code which meant that
the first element of each radio button set was being selected
rather than the checked one. This has now been fixed.

2015's sponsorship information has now been added, although
placeholder text is in place on the Sponsorship tab. 2015's
branding has also been added.