Removed spaces in the call to Google
For some reason Google introduced non-breaking spaces into the
values for some of the options. This meant that it wasn't accepting
ajax calls without them in. I therefore decided to remove the spaces
all together from the Google spreadsheet, and had to do some hacking
to get this to work with the current JavaScript.