Added About page and make the navbar code reusable
When I added the About page I realised that having separate
navbar code in each HTML file was a horrible waste of time
as I had to go into each file each time I wanted to modify
the navbar. Therefore, the navbar.js javascript replaces a
div with the class navbar_replacement with the navbar code
contained in the script. This means that changes to the navbar
code in a single file is seen automatically across the entire
site.